Extract settings persistence engine into UserSettingsStore
Split vtsearch/settings.py (~1935 lines) by moving the lock-ordering-
sensitive engine (cross-process file locking, two-tier caches, the
one-shot legacy migration, and the bidirectional sync state machine)
into a new vtsearch/settings_store.py as UserSettingsStore.

settings.py keeps the schema/policy layer (Pydantic-driven accessor
generation, tier routing, CLI fallbacks, effective-value resolvers) and
delegates the engine work to a module-level store instance. The shared
mutable containers (_settings_lock, _user_caches, _sync_state, _syncing)
stay as module globals and are passed by reference into the store so
external importers (vtsearch.achievements, the sync-source tests) and
the store mutate one set of objects. Behavior is unchanged.
