+++ vtscore/detectors/store.py
@@ -57,7 +57,7 @@ def _write_detector(path: Path, data: dict) -> None:
     # overwrite the same detector file can't truncate each other's in-flight
     # tmp file or chase one that was already renamed away (which surfaced as
     # ``FileNotFoundError: '<name>.json.tmp' -> '<name>.json'`` from
-    # ``os.replace``).  Mirrors ``vtsearch.settings._atomic_write`` and
+    # ``os.replace``).  Mirrors ``vtsearch.settings_store._atomic_write`` and
     # ``vtscore.io.atomic_write_text``.
     tmp = path.with_name(f"{path.name}.{os.getpid()}.{uuid.uuid4().hex}.tmp")
     try:
