+++ vtscore/io.py
@@ -81,7 +81,7 @@ def atomic_write_text(path: Path | str, text: str) -> None:
     # Per-writer unique tmp suffix so two threads (or two processes)
     # racing to overwrite the same destination can't truncate each
     # other's in-flight tmp file or chase one that was already renamed
-    # away.  Mirrors the pattern in ``vtsearch.settings._atomic_write``.
+    # away.  Mirrors the pattern in ``vtsearch.settings_store._atomic_write``.
     tmp = p.with_name(f"{p.name}.{os.getpid()}.{uuid.uuid4().hex}.tmp")
     try:
         with open(tmp, "w", encoding="utf-8", newline="") as f:
